+++ clang/lib/AST/ExprConstant.cpp
@@ -2576,6 +2576,7 @@ static bool HandleConversionToBool(const APValue &Val, bool &Result) {
   switch (Val.getKind()) {
   case APValue::None:
   case APValue::Indeterminate:
+  case APValue::Erroneous:
     return false;
   case APValue::Int:
     Result = Val.getInt().getBoolValue();
@@ -4168,10 +4169,10 @@ findSubobject(EvalInfo &Info, const Expr *E, const CompleteObject &Obj,
 
   // Walk the designator's path to find the subobject.
   for (unsigned I = 0, N = Sub.Entries.size(); /**/; ++I) {
-    // Reading an indeterminate value is undefined, but assigning over one is OK.
+    // Reading an indeterminate value is undefined, but assigning over one is
+    // OK.
     if ((O->isAbsent() && !(handler.AccessKind == AK_Construct && I == N)) ||
-        (O->isIndeterminate() &&
-         !isValidIndeterminateAccess(handler.AccessKind))) {
+        (O->isUninit() && !isValidIndeterminateAccess(handler.AccessKind))) {
       // Object has ended lifetime.
       // If I is non-zero, some subobject (member or array element) of a
       // complete object has ended its lifetime, so this is valid for
@@ -4180,8 +4181,7 @@ findSubobject(EvalInfo &Info, const Expr *E, const CompleteObject &Obj,
         return false;
       if (!Info.checkingPotentialConstantExpression()) {
         Info.FFDiag(E, diag::note_constexpr_access_uninit)
-            << handler.AccessKind << O->isIndeterminate()
-            << E->getSourceRange();
+            << handler.AccessKind << O->isUninit() << E->getSourceRange();
         NoteLValueLocation(Info, Obj.Base);
       }
       return handler.failed();
@@ -4990,6 +4990,7 @@ struct CompoundAssignSubobjectHandler {
     case APValue::Vector:
       return foundVector(Subobj, SubobjType);
     case APValue::Indeterminate:
+    case APValue::Erroneous:
       Info.FFDiag(E, diag::note_constexpr_access_uninit)
           << /*read of=*/0 << /*uninitialized object=*/1
           << E->getLHS()->getSourceRange();
@@ -7768,6 +7769,7 @@ class APValueToBufferConverter {
     // Dig through Src to find the byte at SrcOffset.
     switch (Val.getKind()) {
     case APValue::Indeterminate:
+    case APValue::Erroneous:
     case APValue::None:
       return true;