+++ clang/lib/CodeGen/CodeGenFunction.h
@@ -5532,8 +5532,9 @@ private:
                                         unsigned Type,
                                         llvm::IntegerType *ResType);
 
-  void emitZeroOrPatternForAutoVarInit(QualType type, const VarDecl &D,
-                                       Address Loc);
+  void
+  emitZeroOrPatternForAutoVarInit(QualType type, const VarDecl &D, Address Loc,
+                                  LangOptions::TrivialAutoVarInitKind Kind);
 
 public:
   enum class EvaluationOrder {